requirements.txt
adbench==0.1.11
numpy==1.24.3
catboost<1.2
mysql-connector-python~=8.1.0